Upgrade to jaq 1
