+++ crates/cli/Cargo.toml
@@ -31,10 +31,13 @@ dashmap = "5.4.0"
 dirs = "5.0.0"
 futures = "0.3.29"
 humantime = "2.1.0"
-indexmap = "1.9.3"
+indexmap = "2.2.6" # needs to be in sync with jaq's
 is-terminal = "0.4.4"
 jaq-core = "1.2.1"
+jaq-interpret = "1.2.1"
+jaq-parse = "1.0.2"
 jaq-std = "1.2.1"
+jaq-syn = "1.1.0"
 notify-rust = "4.9.0"
 once_cell = "1.17.1"
 serde_json = "1.0.107"
